requirements.txt
pytube>=15.0.0
pandas==2.2.1
yt-dlp==2024.3.10
flask==3.0.2
transformers==4.38.2
torch==2.2.1
sentencepiece==0.2.0
python-dotenv==1.0.1
accelerate==0.27.2 